+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: mapter <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/23 23:20:06 by mapter            #+#    #+#             */
-/*   Updated: 2023/02/21 22:11:25 by mapter           ###   ########.fr       */
+/*   Updated: 2023/02/24 16:42:39 by mapter           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	*ft_memset(void *s, int c, size_t n)
 {
 	char	*ptr;
-	int		i;
+	size_t	i;
 
 	ptr = (char *) s;
 	i = 0;
